Update README.md
We are looking to change the project description in a way that strongly suggests the applicant not use any CSS frameworks. We have had several submissions where it was difficult to gauge the CSS skill level, so this will help to make that more clear.

In addition, we are open to changing other requirements in the interest of time. Writing custom css could take more time, so we would encourage the removal of the "average selection time" component on the final page. That will reduce the complexity of the javascript, and therefore reduce the time needed to complete it.